rename to node-contacts
