--- src/controllers/contacts.controller.js
@@ -1,16 +1,0 @@
-'use strict';
-
-const nodeContactModel = require('~models/node-contact.model');
-
-const createContacts = (request, response) => {
-  const contacts = request.body.contacts;
-
-  nodeContactModel
-    .create(contacts)
-    .then(result => response.send(result))
-    .catch(err => res.status(500).send(errorResponse(500, err.message)));
-};
-
-module.exports = {
-  createContacts
-};
